owsa_opt_strat updates
Added parameter list option to owsa_opt_start(). Closes #66

Parameter ordering in plot matches the input by the user in the owsa_opt_strat() function. If no input is provided, all parameters are shown and the order matches the order from the owsa object (rather than being alphabetical)

Added a new parameter in owsa_plot_strat and add_common_aes that allows for the text size of the facet labels to be set manually. Leaving this blank defaults to facet label size of the previous build.

As for the middle two issues from #67 , it seems that the spacing and number of the tick marks on the x-axes are ultimately controlled by the pretty() function. The n_x_ticks parameter provides pretty() with a target number of tick marks, but it seems like pretty() makes some internal decisions about what will look best. I will keep messing around with the parameter values to see if something looks better.

The tick marks don't line up with the rectangles because pretty() is forcing the tick marks to appear at nice, round numbers that frame the region of interest. Should this be changed so that tick marks show the values of the exact borders?