number_ticks() fix
the parameter 'n' in number_ticks() is supposed to supply the desired number of tick marks on the x-axes, but in the function pretty(), which 'n' is ultimately passed to, this parameter sets the target number of intervals. 'n' was changed to 'n+1' within pretty() in the number_ticks() function.