Implement special case of calculate_outcomes()
calc_evppi and calc_evsi now use an adjusted version of calculate_outcomes to remain consistent with Fernando's coding suggestions in "A Gaussian Approximation Approach for Value of Information Analysis".  These functions still do not seem to be working correctly, but this seems like a step in the right direction.