initial evsi updates for evsi vs. n plots
evsi output is long form data.frame with WTP, n, and EVSI
for plot.evsi to show how evsi changes with new sample size ('n'), the way the function handles n needs to be changed.
For these plots, n must be applied to all parameters used in evsi, but in original implementation n was allowed to be specified for each parameter in the model
function needs to be updated and clarified to maintain original functionality while streamlining the plotting functionality.