Metamodel failure warnings/errors
for multiway metamodels, k in the interaction terms must be at least 3.  If k is less than 3, the k values will be set automatically, which will almost always result in a model that fails to fit due to memory constraints when n_params > 3.
Added another warning when ((k - 1) ^ n_params > 500). This is a somewhat arbitrary cutoff, but I found that models that exceed this cutoff tend to fail to fit or take too long.  The mgcv fitting process constructs several (k - 1) ^ n_params by (k - 1) ^ n_params matrices, which rapidly become too large.